==> NOTEBOOKS/00_PREPROCESSING/00_PREPROCESSING_DROPNA.ipynb <==
## PREPROCESSING OF NA AND 0 VALUES

# Import Libs

import pandas as pd
from ASSETS.CLASS.Viz import Viz

# Open File

df = pd.read_csv('../../DATA/train.csv', nrows=2000000)
df.head()

## Size of the Dataset

df.shape

## Data Type and Nan count by features

df.info()

df.isna().sum()

# Drop the rows without data

df = df.dropna()
df.shape

## First Informations about Dataset

df.describe(include='all')

## check if we have values = 0

df.loc[(df['pickup_longitude'] == 0.0) | (df['pickup_latitude'] == 0.0) | (df['dropoff_longitude'] == 0.000000)  | (df['dropoff_latitude'] == 0.000000) | (df['fare_amount'] == 0.0) | (df['passenger_count'] == 0)]

# DELETE ALL ROWS WITH 0 VALUES + DELETE ALL ROWS WITH FARE AMOUNT UNDER 0$


df.drop(df.loc[(df['pickup_longitude'] == 0.0) | (df['pickup_latitude'] == 0.0) | (df['dropoff_longitude'] == 0.000000)  | (df['dropoff_latitude'] == 0.000000) | (df['fare_amount'] <= 1) | (df['passenger_count'] == 0)].index, inplace=True)
df.head()

# DISPLAY NUMBER OF ROW AFTER PREPROCESSING

df.shape

# EXPORT DATAFRAME TO CSV FOR DATA VIZUALISATION

df.to_csv('../../DATA/PREPROCESSING/prepro_00_dropna.csv', index=False)
