==> NOTEBOOKS/00_PREPROCESSING/01b_PREPROCESSING_RESULT_GMAP_API.ipynb <==
# PREPROCESSING OF RESULTS DRIVING DISTANCE AND DURATION FROM GMAP API

# IMPORT LIBS

import pandas as pd
import plotly.express as px

# OPEN DATASET

df = pd.read_csv('../../DATA/PREPROCESSING/prepro_01a_geo_gmap.csv')
df.head()

# SCATTER PLOT OF DISTANCE AND DRIVING DISTANCE

fig = px.scatter(df, x='Distance (miles)', y='fare_amount')
fig.show()

fig = px.scatter(df, x='Driving distance (miles)', y='fare_amount')
fig.show()
