==> NOTEBOOKS/00_PREPROCESSING/01c_PREPROCESSING_GEOGRAPHIC_CALCUL.ipynb <==
# DRIVING DISTANCE AND DURATION CALCUL

# IMPORT LIBS

import pandas as pd
import plotly.express as px

# OPEN DATASET

df = pd.read_csv('../../DATA/PREPROCESSING/prepro_01_geo.csv')
df

# CALCUL OF DISTANCE WITH CAR

# Calculer la distance en voiture en utilisant une approximation
df['Driving distance (miles)'] = df['Distance (miles)'] * 1.5  # Facteur de vitesse moyen pour les routes urbaines

# Combinaison pondérée de la distance à vol d'oiseau et de la distance en voiture
df['Driving distance (miles)'] = 0.8 * df['Distance (miles)'] + 0.2 * df['Driving distance (miles)']

df.head()

# CALCUL OF DURATION TIME

#  la vitesse moyenne des voitures dans les rues de New York varie de 15 à 25 km/h en heure de pointe. on choisie une valeur moyenne, par exemple 20 km/h

speed = 20  # Vitesse moyenne dans New York km/h
df['Driving duration (min)'] = df['Driving distance (miles)'] / (speed / 60)

# Afficher le DataFrame avec les temps de trajet calculés
df

# EXPORT DF

df.to_csv('../../DATA/PREPROCESSING/prepro_01c_geo_calcul.csv', index=False)
