==> NOTEBOOKS/00_PREPROCESSING/01d_PREPROCESSING_RESULT_CALCUL_DIST.ipynb <==
# PREPROCESSING OF RESULTS DRIVING DISTANCE AND DURATION FROM CALCUL

# IMPORT LIBS

import pandas as pd
import plotly.express as px

# OPEN DATASET

df = pd.read_csv('../../DATA/PREPROCESSING/prepro_01c_geo_calcul.csv')
df.head()

# SCATTER PLOT OF DISTANCE AND DRIVING DISTANCE

# fig = px.violin(df, y="Driving distance (miles)", box=True,
#                 # points="all",
#                 # hover_data=df.columns,
#                 title='Violin plot of Driving distance distribution'
#                 )
# fig.show()

# DELETE ALL RACE UNDER 0.5 MILES

df = df.loc[(df['Driving distance (miles)'] > 0.5)]

# DISPLAY NUMBER OF ROWS AFTER PREPROCESSING

df.shape

# EXPORT CSV

df.to_csv('../../DATA/PREPROCESSING/prepro_01d_geo_result_calcul.csv', index=False)
