==> NOTEBOOKS/00_PREPROCESSING/02_PREPROCESSING_FARE_AMOUNT.ipynb <==
# PREPROCESSING OF FARE AMOUNT

# IMPORT LIBS

import pandas as pd
import plotly.express as px
import numpy as np

# OPEN DATASET

df = pd.read_csv('../../DATA/PREPROCESSING/prepro_01d_geo_result_calcul.csv')
df.head()

# VISUALISATION OF OUTLIERS

## DISPLAY FARE AMOUNT DISTRIBUTION

# fig = px.box(df, y="fare_amount", title='Box plot of fare amount distribution')
# fig.show()

We can saw the box plot is crushed, we gonna try to display a violin plot to saw the density of outliers

# fig = px.violin(df, y="fare_amount", box=True, # draw box plot inside the violin
#                 # points='all',# can be 'outliers', or False
#                 title='Violin plot of fare amount distribution'
#                 )
# fig.show()

### Calcul D’Outliers :
X < q1 - 1.5 x iqr X > q3 - 1.5 x iqr

q1 = df['fare_amount'].quantile([.25])
q1 = q1[0.25]
q3 = df['fare_amount'].quantile([.75])
q3 = q3[0.75]

iqr = q3 - q1

print(f" IQR = {iqr}, q1 = {q1}, q3 = {q3}")

outliers_superieur = q3 + (1.5 * iqr)
outliers_inferieur = q1 - (1.5 * iqr)

print(f'Outliers Inférieur = {outliers_inferieur}')
print(f'Outliers Supérieur = {outliers_superieur}')

## DELETE OUTLIERS

df = df.loc[(df['fare_amount'] <= outliers_superieur) & (df['fare_amount'] >= outliers_inferieur)]
df.shape

# DISPLAY A NEW VIOLIN PLOT OF DISTRIBUTION OF FARE AMOUNT WITHOUT OUTLIERS

# fig = px.violin(df, y="fare_amount", box=True, # draw box plot inside the violin
#                 # points='all',# can be 'outliers', or False
#                 title='Violin plot of fare amount distribution'
#                 )
# fig.show()

## DISPLAY PASSENGERS DISTRIBUTION

# fig = px.violin(df, y="fare_amount", x="passenger_count", box=True,
#                 # points="all",
#                 # hover_data=df.columns,
#                 title='Violin plot of numbers of passengers count distribution'
#                 )
# fig.show()

With this plot, we saw the passengers count didn't impact us fare amount, so we don't gonnawork with this feature

# DISPLAY ROWS NUMBERS AFTER PREPROCESSING

df.shape

# EXPORT CSV

df.to_csv('../../DATA/PREPROCESSING/prepro_02_fare_amount.csv', index=False)
