==> NOTEBOOKS/00_PREPROCESSING/03_PREPROCESSING_CALENDAR.ipynb <==
# PREPROCESSING OF TIME

# IMPORT LIBS

import pandas as pd
import holidays
from datetime import datetime
import plotly.express as px

# OPEN DATASET

df = pd.read_csv('../../DATA/PREPROCESSING/prepro_02_fare_amount.csv')
df.head()

# WHAT IS THE DAY OF THE RACE

df['day_of_the_week']= df['key'].apply(lambda x: pd.Timestamp(x).day_name())
df.head()

# THIS DAY IS PUBLIC HOLIDAY ?

## TRANSLATE DATE FORMAT AND CREATE FEATURE DATE AND TIME

df["key"] = pd.to_datetime(df["key"])
df["date"] = df["key"].dt.date
df["time"] = df["key"].dt.time

df.head()

#  GET IF THE DAY IS PUBLIC HOLIDAY BY HOLIDAYS LIB

ny_holidays = holidays.US(state='NY')

# Vérifier si chaque date est un jour férié
df["public_holiday"] = df["date"].apply(lambda x: x in ny_holidays)

df.head()

## DISPLAY THE DISTRIBUTION OF PUBLIC HOLIDAYS IN US DATASET

holidays_value_count = df['public_holiday'].value_counts()

fig = px.pie(values=holidays_value_count.values, labels=holidays_value_count.index)
fig.show()

# DELETE THE KEY AND PICKUP_DATETIME FEATURES

df = df.drop(columns=['key', 'pickup_datetime'])
df.head()

# EXPORT CSV

df.to_csv('../../DATA/PREPROCESSING/prepro_03_calendar.csv', index=False)
