==> NOTEBOOKS/00_PREPROCESSING/04_PREPROCESSING_PRICE_BY_MILES.ipynb <==
# PREPROCESSING OF PRICE BY MILES

# IMPORT LIBS

import pandas as pd
import plotly.express as px

# OPEN DATASET

df = pd.read_csv('../../DATA/PREPROCESSING/prepro_03_calendar.csv')
df.head()

# CREATE NEW FEATURE PRICE BY MILES

df['price_by_miles'] = df['fare_amount'] / df['Driving distance (miles)']
df.head()

# DISPLAY A NEW VIOLIN PLOT OF DISTRIBUTION OF PRICE BY MILES

# fig = px.violin(df, y="price_by_miles", box=True, # draw box plot inside the violin
#                       # points='all',# can be 'outliers', or False
#                       title='Violin plot of price_by_km distribution'
#                       )
# fig.show()

# OUTLIERS

### Calcul D’Outliers :
X < q1 - 1.5 x iqr X > q3 - 1.5 x iqr

q1 = df['price_by_miles'].quantile([.25])
q1 = q1[0.25]
q3 = df['price_by_miles'].quantile([.75])
q3 = q3[0.75]

iqr = q3 - q1

print(f" IQR = {iqr}, q1 = {q1}, q3 = {q3}")

outliers_superieur = q3 + (1.5 * iqr)
outliers_inferieur = q1 - (1.5 * iqr)

print(f'Outliers Inférieur = {outliers_inferieur}')
print(f'Outliers Supérieur = {outliers_superieur}')

## DELETE OUTLIERS

df = df.loc[(df['price_by_miles'] <= outliers_superieur) & (df['price_by_miles'] >= outliers_inferieur)]
df.shape

# DISPLAY A NEW VIOLIN PLOT OF DISTRIBUTION OF PRICE BY MILES WITHOUT OUTLIERS

fig = px.violin(df, y="price_by_miles", box=True, # draw box plot inside the violin
                # points='all',# can be 'outliers', or False
                title='Violin plot of price by miles distribution'
                )
fig.show()

# EXPORT CSV

df.to_csv('../../DATA/PREPROCESSING/prepro_04_price_by_miles.csv', index=False)
