==> NOTEBOOKS/00_PREPROCESSING/05_PREPROCESSING_NIGHT_OR_DAY.ipynb <==
# PREPROCESSING OF TIME OF THE DAY

# IMPORT LIBS

import pandas as pd
import plotly.express as px

# OPEN DATASET

df = pd.read_csv('../../DATA/PREPROCESSING/prepro_04a_price_by_time.csv')
df.head()

df.dtypes

# CREATE NEW FEATURE DAY OR NIGHT

## TRANSFORM TIME FORMAT TO DATETIME

df["time"] = pd.to_datetime(df["time"], format='%H:%M:%S')
df.head()

## CREATION DE LA FEATURE DAY OR NIGHT

# Créer une fonction lambda pour déterminer si la course a eu lieu pendant la journée ou la nuit
def day_or_night(time):
    heure_min = time.hour * 60 + time.minute
    # heure jour : 7 à 22h
    if heure_min >= 7 * 60 and heure_min < 21 * 60:
        return True
    else:
        return False

# Appliquer la fonction lambda à chaque élément de la fonctionnalité "heure" pour créer la fonctionnalité "jour_nuit"
df["daytime"] = df["time"].apply(lambda x: day_or_night(x))
df.head()

# EXPORT CSV

df.to_csv('../../DATA/PREPROCESSING/prepro_05_DAY_OR_NIGHT.csv', index=False)
