==> NOTEBOOKS/00_PREPROCESSING/06_PREPROCESSING_ML_DATASET.ipynb <==
# PREPROCESSING TO PREPARE THE DATASET FOR MODELS

# IMPORT LIBS

import pandas as pd

# OPEN DATASET

df = pd.read_csv('../../DATA/PREPROCESSING/prepro_05_DAY_OR_NIGHT.csv')
df.head()

# DROP USELESS COLUMN

df = df.drop(columns=['pickup_longitude', 'pickup_latitude', 'dropoff_longitude', 'dropoff_latitude', 'date', 'time', 'price_by_miles', 'price_by_time'], axis=1)
df.head()

# GET DUMMIES DATASET

df = pd.get_dummies(df)
df.head()

# EXPORT CSV

df.to_csv('../../DATA/PREPROCESSING/prepro_06_ML_DATASET.csv', index=False)
